+++ packages/graphmarket-entities/src/index.ts
@@ -1,4 +1,2 @@
-export { default as Product } from './Product';
-export * from './Product';
-export { default as User } from './User';
-export * from './User';
+export * from './product';
+export * from './user';
