+++ packages/graphmarket-entities/src/product/ProductCategories.ts
@@ -1,0 +1,41 @@
+import { registerEnumType } from 'type-graphql';
+
+/**
+ * Product categories.
+ */
+enum ProductCategories {
+  AUTOMOTIVE_AND_POWERSPORTS = 'AUTOMOTIVE_AND_POWERSPORTS',
+  BABY_PRODUCTS = 'BABY_PRODUCTS',
+  BEAUTY = 'BEAUTY',
+  BOOKS = 'BOOKS',
+  CAMERA_AND_PHOTO = 'CAMERA_AND_PHOTO',
+  CELL_PHONES_AND_ACCESSORIES = 'CELL_PHONES_AND_ACCESSORIES',
+  COLLECTIBLE_COINS = 'COLLECTIBLE_COINS',
+  CONSUMER_ELECTRONICS = 'CONSUMER_ELECTRONICS',
+  ENTERTAINMENT_COLLECTIBLES = 'ENTERTAINMENT_COLLECTIBLES',
+  FINE_ART = 'FINE_ART',
+  GROCERY_AND_GOURMET_FOOD = 'GROCERY_AND_GOURMET_FOOD',
+  HEALTH_AND_PERSONAL_CARE = 'HEALTH_AND_PERSONAL_CARE',
+  HOME_AND_GARDEN = 'HOME_AND_GARDEN',
+  INDEPENDENT_DESIGN = 'INDEPENDENT_DESIGN',
+  INDUSTRIAL_AND_SCIENTIFIC = 'INDUSTRIAL_AND_SCIENTIFIC',
+  MAJOR_APPLIANCES = 'MAJOR_APPLIANCES',
+  MUSIC = 'MUSIC',
+  MUSICAL_INSTRUMENTS = 'MUSICAL_INSTRUMENTS',
+  OFFICE_PRODUCTS = 'OFFICE_PRODUCTS',
+  OUTDOORS = 'OUTDOORS',
+  PERSONAL_COMPUTERS = 'PERSONAL_COMPUTERS',
+  PET_SUPPLIES = 'PET_SUPPLIES',
+  SOFTWARE = 'SOFTWARE',
+  SPORTS = 'SPORTS',
+  SPORTS_COLLECTIBLES = 'SPORTS_COLLECTIBLES',
+  TOOLS_AND_HOME_IMPROVEMENTS = 'TOOLS_AND_HOME_IMPROVEMENTS',
+  TOYS_AND_GAMES = 'TOYS_AND_GAMES',
+  VIDEO_AND_DVD_AND_BLU_RAY = 'VIDEO_AND_DVD_AND_BLU_RAY',
+  VIDEO_GAMES = 'VIDEO_GAMES',
+  WATCHES = 'WATCHES',
+}
+
+registerEnumType(ProductCategories, { name: 'ProductCategories' });
+
+export default ProductCategories;
