+++ packages/graphmarket-entities/src/product/index.ts
@@ -1,0 +1,3 @@
+export { default as Product } from './Product';
+export { default as ProductCategories } from './ProductCategories';
+export { default as ProductConditions } from './ProductConditions';
