+++ packages/graphmarket-entities/src/user/index.ts
@@ -1,0 +1,4 @@
+export { default as User } from './User';
+export { default as UserExternal } from './UserExternal';
+export { default as UserGenders } from './UserGenders';
+export { default as UserRoles } from './UserRoles';
