+++ packages/graphmarket-service-products/src/resolvers/index.ts
@@ -1,2 +1,3 @@
 export { default as ProductResolver } from './ProductResolver';
 export { default as resolveProductReference } from './resolveProductReference';
+export { default as UserProductResolver } from './UserProductResolver';
